backend/requirements.txt
fastapi
uvicorn
playwright
beautifulsoup4
lxml
requests
python-dotenv
